+++ python/pyspark/sql/tests/test_conversion.py
@@ -20,6 +20,7 @@ from zoneinfo import ZoneInfo
 
 from pyspark.errors import PySparkValueError
 from pyspark.sql.conversion import (
+    ArrowArrayToPandasConversion,
     ArrowTableToRowsConversion,
     LocalDataToArrowConversion,
     ArrowArrayConversion,
@@ -39,7 +40,7 @@ from pyspark.sql.types import (
     StructType,
     UserDefinedType,
 )
-from pyspark.testing.objects import ExamplePoint, ExamplePointUDT
+from pyspark.testing.objects import ExamplePoint, ExamplePointUDT, PythonOnlyPoint, PythonOnlyUDT
 from pyspark.testing.utils import have_pyarrow, pyarrow_requirement_message
 
 
@@ -374,6 +375,46 @@ class ConversionTests(unittest.TestCase):
         ]:
             output = ArrowArrayConversion.localize_tz(arr)
             self.assertEqual(output, expected, f"{output.tolist()} != {expected.tolist()}")
+
+
+@unittest.skipIf(not have_pyarrow, pyarrow_requirement_message)
+class ArrowArrayToPandasConversionTests(unittest.TestCase):
+    def test_udt_convert_numpy(self):
+        import pyarrow as pa
+
+        udt = ExamplePointUDT()
+
+        # basic conversion with nulls
+        arr = pa.array([[1.0, 2.0], None, [3.0, 4.0]], type=pa.list_(pa.float64()))
+        result = ArrowArrayToPandasConversion.convert_numpy(arr, udt, ser_name="my_point")
+        self.assertIsInstance(result.iloc[0], ExamplePoint)
+        self.assertEqual(result.iloc[0], ExamplePoint(1.0, 2.0))
+        self.assertIsNone(result.iloc[1])
+        self.assertEqual(result.iloc[2], ExamplePoint(3.0, 4.0))
+        self.assertEqual(result.name, "my_point")
+
+        # empty
+        result = ArrowArrayToPandasConversion.convert_numpy(
+            pa.array([], type=pa.list_(pa.float64())), udt
+        )
+        self.assertEqual(len(result), 0)
+
+        # PythonOnlyUDT
+        result = ArrowArrayToPandasConversion.convert_numpy(
+            pa.array([[5.0, 6.0]], type=pa.list_(pa.float64())), PythonOnlyUDT()
+        )
+        self.assertIsInstance(result.iloc[0], PythonOnlyPoint)
+        self.assertEqual(result.iloc[0], PythonOnlyPoint(5.0, 6.0))
+
+    def test_udt_chunked_array(self):
+        import pyarrow as pa
+
+        chunk1 = pa.array([[1.0, 2.0]], type=pa.list_(pa.float64()))
+        chunk2 = pa.array([[3.0, 4.0]], type=pa.list_(pa.float64()))
+        chunked = pa.chunked_array([chunk1, chunk2])
+        result = ArrowArrayToPandasConversion.convert_numpy(chunked, ExamplePointUDT())
+        self.assertEqual(result.iloc[0], ExamplePoint(1.0, 2.0))
+        self.assertEqual(result.iloc[1], ExamplePoint(3.0, 4.0))
 
 
 if __name__ == "__main__":
